Move gate_matrices.py to simulation.matrices
