+++ pyquil/tests/test_unitary_tools.py
@@ -2,10 +2,12 @@ import numpy as np
 import pytest
 
 from pyquil import Program
-from pyquil import gate_matrices as mat
-from pyquil.gates import CCNOT, CNOT, CZ, H, MEASURE, PHASE, RX, RY, RZ, X, Y, Z
 from pyquil.experiment import plusX, minusZ
+from pyquil.gates import CCNOT, CNOT, CZ, H, MEASURE, PHASE, RX, RY, RZ, X, Y, Z
 from pyquil.paulis import sX, sY, sZ
+from pyquil.quilatom import MemoryReference, Parameter
+from pyquil.quilbase import Declare
+from pyquil.simulation import matrices as mat
 from pyquil.unitary_tools import (
     qubit_adjacent_lifted_gate,
     program_unitary,
@@ -14,8 +16,6 @@ from pyquil.unitary_tools import (
     lifted_pauli,
     lifted_state_operator,
 )
-from pyquil.quilatom import MemoryReference, Parameter
-from pyquil.quilbase import Declare
 
 
 def test_random_gates():
